refactor: shift context constants into 4.0
