+++ src/4.0/wrap.ts
@@ -1,6 +1,6 @@
 import { hashToBuffer, isStringArray } from "../shared/utils";
 import { MerkleTree } from "../shared/merkle";
-import { ContextType, ContextUrl } from "../shared/@types/document";
+import { ContextUrl, ContextType } from "./validate/context";
 import { NoExtraProperties, V4Document, V4WrappedDocument, W3cVerifiableCredential } from "./types";
 import { digestCredential } from "../4.0/digest";
 import { encodeSalt, salt } from "./salt";
